Refactor nav section styles
