+++ src/components/Nav/Container.jsx
@@ -1,0 +1,17 @@
+import styled from "styled-components"
+
+export const Container = styled.nav`
+  z-index: 2;
+  height: 7rem;
+  background-color: white;
+  border-bottom: 2px solid #f4f4f4;
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+`
+
+export default Container
