+++ src/components/Nav/Icon.jsx
@@ -1,0 +1,7 @@
+import styled from "styled-components"
+
+export const Icon = styled.img`
+  height: 8rem;
+`
+
+export default Icon
