+++ src/components/Nav/IconWrapper.jsx
@@ -1,0 +1,9 @@
+import styled from "styled-components"
+
+export const IconWrapper = styled.div`
+  cursor: pointer;
+  margin-left: 1.2rem;
+  margin-top: -4px;
+`
+
+export default IconWrapper
