+++ src/pages/index.js
@@ -1,6 +1,6 @@
 import React, { useRef } from "react"
 import SEO from "../components/seo"
-import Nav from "../components/nav"
+import Nav from "../components/Nav/Nav"
 import Hero from "../components/hero"
 import About from "../components/About/About"
 import Sponsor from "../components/sponsor"
